+++ Assets/PlayScene/C#Script/Enemy/Enum/EnemyState.cs
@@ -7,7 +7,8 @@ public class EnemyState : MonoBehaviour
     public enum EscapeEnemyState
     {
         Search,
-        Dush
+        Dush,
+        Die
     }
 
 }
