+++ Assets/PlayScene/Resouce/Enemy/Model/ty.prefab
@@ -257,8 +257,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   DefeatTag: Player
   mPursuer: {fileID: 0}
-  mDesiredPower: 15
-  mForwardSearch: {fileID: 0}
+  mRithtTurnPower: 1.5
+  mForwardSearch: {fileID: 6424683430257139420}
 --- !u!195 &85790663884643903
 NavMeshAgent:
   m_ObjectHideFlags: 0
@@ -269,7 +269,7 @@ NavMeshAgent:
   m_Enabled: 1
   m_AgentTypeID: 0
   m_Radius: 0.5
-  m_Speed: 3.5
+  m_Speed: 12
   m_Acceleration: 8
   avoidancePriority: 50
   m_AngularSpeed: 120
@@ -2435,10 +2435,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 61eb061ec33646f41863de04e47378a9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  mTargetName: 
-  mSearchRange: 100
-  mRayNum: 3
-  mIsDetectionGround: 0
+  mOddRayNum: 9
+  mRayDeg: 15
+  mDownPercentage: 2
+  mSearchRange: 10
 --- !u!1 &8017328756027604814
 GameObject:
   m_ObjectHideFlags: 0
