--- Assets/YanoTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b791948e30a85f244bfaa803ac4c4412
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
