--- Assets/YanoTest/Navigation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9897a05a6cb7f704baf1399dea1203c5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
