--- Assets/YanoTest/Navigation.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: be9f60124d3e8ab47ba99131a9b1af90
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
