--- Assets/YanoTest/Navigation/NavMesh.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f77daca480647ef4495bebfbe5c76ee7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
